FIX: Fix bug on MacOS platform where merged arrays couldn't be converted back to volume
* Fix conversion problems between Slicer and Numpy. (Slicer expects int32 and not int64)

+++ RVXLiverSegmentation/RVXLiverSegmentationLib/ExtractVesselStrategies.py
@@ -177,8 +177,8 @@ def mergeVolumes(volumes, volName):
   -------
   Tuple[vtkMRMLVolumeNode, vtkMRMLModelNode]
   """
-  # Extract list of volumes as list of np arrays
-  npVolumes = [slicer.util.arrayFromVolume(volume).astype(int) for volume in volumes]
+  # Extract list of volumes as list of np arrays in int32 format
+  npVolumes = [slicer.util.arrayFromVolume(volume).astype("int32") for volume in volumes]
 
   # Merge all volumes in one
   mergedVol = npVolumes[0]
